fix(judge,ranker): reconcile matmul SizeClass — one shared aspect-carrying key on producer + consumer
Fixes a latent correctness bug (found by slice 2): the Judge measured
latency for ANY non-square matmul was unreachable or poisoned. SizeClass
was an aspect-blind log2(total_elements); the PRODUCER keyed matmul on
m*n (output) while the ranker CONSUMER keyed on m*k (LHS input) — square
matmuls agreed by accident, every rectangular GEMM (attention scores,
FFN projections, decode GEMVs) disagreed -> the ranker never found the
producer's entry (fell to Layer-1) or hit a colliding cell (the oracle
keeps MIN latency across a collision -> poison). This affected all
models, not just decode; it's slice 4's prerequisite for costing the
decode path's non-square attention matmuls.

The fix = a SINGLE shared key derivation both sides call from the operand
shapes (fuel-ir/src/dispatch.rs):
- SizeClass widened u8 -> u32 (design (b): keeps ~120 SizeClass(N)
  literals compiling via integer inference — an enum (a) would have
  forced ~120 edits across 4 crates incl. correctness-fixtures).
- SizeClass::matmul(m,n,k) packs (log2(m*n)<<16)|(log2(m)<<8)|log2(k):
  the high byte keeps pick_nearest magnitude-ordered by output size, the
  low bytes carry aspect so a GEMV never collides with a same-output
  square. SizeClass::for_op(op, shapes) is the consumer entry: matmul
  reads (m,n,k) from lhs=[..,m,k]/rhs=[..,k,n] (trailing dims,
  batch-invariant), everything else -> from_elem_count (unchanged Total
  key).
- Reconciled: producer judge/mod.rs run(); consumers cost.rs, plan.rs
  stamp, route_picker.rs -> all call the shared helper.
- Un-deferred slice 2's dropped non-square cells (now key correctly):
  FFN-width [1,2048]x[2048,5632] + QK^T [1,64]x[64,128] + attn-output
  [1,128]x[128,64] decode GEMVs (DECODE_HEAD_DIM/DECODE_FFN).

PROFILE_REPORT_VERSION 3 -> 4 (SizeClass repr/serde changed); v3 reports
load as Ok(None) -> clean re-Judge (verified via version_mismatch_yields
_empty_oracle).

Documented follow-ups (bounded this slice to producer + ranker as
specified): scheduling.rs (transitional, graph-less recommend_for_node
can't reach k; superseded by the ranker) + capture.rs (fixture path, not
loaded on the default Judge path) keep their old matmul keying —
compile-fixed (u32), semantically noted in-code.

Born-red (observed): ranker_lookup_hits_producer_decode_gemv_cell —
producer reverted to m*n keying -> FAILED (SizeClass(12) != SizeClass
(786443)=0x0C000B) -> restored -> ok; dispatch_table_built... genuine
producer/consumer-key red->green; 4 fuel-ir unit tests (producer==consumer
for non-square w/ assert_ne old-m*n!=m*k witness; GEMV-vs-same-output-
square distinct; batch-invariant; non-matmul unchanged).

Verified: fuel-dispatch --lib 584/0/1, fuel-core --lib 1312/0/12 (+1),
fuel-ir --lib 4/4, fuel-correctness-fixtures --lib --features capture
33/0.

+++ fuel-core/src/judge/cache.rs
@@ -215,7 +215,7 @@ mod tests {
     use fuel_ir::probe::BackendId;
     use fuel_ir::DType;
 
-    fn entry(backend: BackendId, op: OpKind, size: u8, latency: u64, err: f32) -> ProfileEntry {
+    fn entry(backend: BackendId, op: OpKind, size: u32, latency: u64, err: f32) -> ProfileEntry {
         ProfileEntry {
             op,
             dtype:         DType::F32,
@@ -236,7 +236,7 @@ mod tests {
     fn entry_with_source(
         backend: BackendId,
         op: OpKind,
-        size: u8,
+        size: u32,
         latency: u64,
         err: f32,
         kernel_source: &str,
